add handleEdit
