+++ client/src/components/Home/RightPanel/RightbarWorkoutListItem.jsx
@@ -1,21 +1,56 @@
-import Modal from "react-bootstrap/Modal";
 import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
+import Modal from "react-bootstrap/Modal";
 import { useState } from "react";
 import "./RightbarWorkoutListItem.scss";
 import WorkoutListItemDescription from "./WorkoutListItemDescription";
 import axios from "axios";
+import ModalWorkout from "./ModalWorkout";
+import WorkoutList from "../WorkoutList/WorkoutList";
 
 export default function RightbarWorkoutListItem(props) {
   console.log("in rightbarworkoutlistitem", props);
-  const { title, exercises, id, responseData, setResponseData } = props;
+  const {
+    title,
+    exercises,
+    id,
+    responseData,
+    setResponseData,
+    group_ids,
+  } = props;
   const [fullscreen, setFullscreen] = useState(true);
   const [show, setShow] = useState(false);
-  const [index, setIndex] = useState(0);
 
   function handleShow(breakpoint) {
     setFullscreen(breakpoint);
     setShow(true);
+    return (
+      <ModalWorkout
+        key={id}
+        id={id}
+        title={title}
+        exercises={exercises}
+        responseData={responseData}
+        setResponseData={setResponseData}
+        fullscreen={fullscreen}
+        setFullscreen={setFullscreen}
+        show={show}
+        setShow={setShow}
+      />
+    );
+  }
+
+  function handleEdit(id) {
+    console.log("this is edit", id);
+    return (
+      <WorkoutList
+        key={id}
+        id={id}
+        title={title}
+        exercises={exercises}
+        group_ids={group_ids}
+      />
+    );
   }
 
   function handleDelete(id) {
@@ -33,16 +68,20 @@ export default function RightbarWorkoutListItem(props) {
       });
   }
 
-  function previous() {
-    if (exercises.length < 2) {
-      return;
-    }
-    setIndex((prev) => prev - 1);
-  }
-
-  function next() {
-    setIndex((prev) => prev + 1);
-  }
+  const showModal = (
+    <ModalWorkout
+      key={id}
+      id={id}
+      title={title}
+      exercises={exercises}
+      responseData={responseData}
+      setResponseData={setResponseData}
+      fullscreen={fullscreen}
+      setFullscreen={setFullscreen}
+      show={show}
+      setShow={setShow}
+    />
+  );
 
   return (
     <>
@@ -50,60 +89,32 @@ export default function RightbarWorkoutListItem(props) {
         <Card.Header>Featured</Card.Header>
         <Card.Body>
           <Card.Title>{title.toUpperCase()}</Card.Title>
-          <Card.Text>
-            {exercises.map((item) => (
-              <WorkoutListItemDescription
-                key={item.id}
-                name={item.name}
-                set={item.set}
-                reps={item.reps}
-              />
-            ))}
-          </Card.Text>
+          {exercises.map((item) => (
+            <WorkoutListItemDescription
+              key={item.id}
+              name={item.name}
+              set={item.set}
+              reps={item.reps}
+            />
+          ))}
           <Button variant="primary" onClick={() => handleShow(true)}>
             Start Workout
+          </Button>
+          <Button variant="primary" onClick={() => handleEdit(id)}>
+            Edit
           </Button>
           <Button variant="primary" onClick={() => handleDelete(id)}>
             Delete
           </Button>
         </Card.Body>
+        <Modal
+          show={show}
+          fullscreen={fullscreen}
+          onHide={() => setShow(false)}
+        >
+          {showModal}
+        </Modal>
       </Card>
-      <Modal show={show} fullscreen={fullscreen} onHide={() => setShow(false)}>
-        <Modal.Header closeButton>
-          <Modal.Title>{title}</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {exercises.length !== index ? (
-            <>
-              {exercises[index].name}
-              <img src={exercises[index].gifUrl} alt={exercises[index].name} />
-              <div>Set: {exercises[index].set}</div>
-              <div>Reps: {exercises[index].reps}</div>
-            </>
-          ) : (
-            <>
-              <div> You did it! </div>
-              <img
-                src="https://media1.giphy.com/media/BqijAlej4RV7O/giphy.gif?cid=ecf05e47nf3oiylzaxxzlk98mzb6t8nskx84noqgwicheqhf&rid=giphy.gif&ct=g"
-                alt="mario-party"
-              />
-            </>
-          )}
-          {index > 0 && (
-            <Button className="me-2" onClick={() => previous()}>
-              Previous
-            </Button>
-          )}
-          {exercises.length !== index && (
-            <Button className="me-2" onClick={() => next()}>
-              Next
-            </Button>
-          )}
-          <Button className="me-2" onClick={() => setIndex(0)}>
-            Restart
-          </Button>
-        </Modal.Body>
-      </Modal>
     </>
   );
 }
