+++ client/src/components/Home/RightPanel/WorkoutListItemDescription.jsx
@@ -1,14 +1,13 @@
-import Card from "react-bootstrap/esm/Card";
 import "./WorkoutListItemDescription.scss";
 
 export default function WorkoutListItemDescription(props) {
   const { name, set, reps } = props;
   return (
-    <section>
+    <div>
       {name}
       <div>
         Set: {set} Reps: {reps}
       </div>
-    </section>
+    </div>
   );
 }
